LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-shorthand-serialization-expected.txt


FAIL grid-template-rows: none,
        grid-template-columns: none,
        grid-template-areas: none,
        grid-lanes-direction: column should be valid. assert_equals: expected (string) "column" but got (undefined) undefined
FAIL grid-template-rows: 10px,
        grid-template-columns: none,
        grid-template-areas: none,
        grid-lanes-direction: column should be valid. assert_equals: expected (string) "column" but got (undefined) undefined
FAIL grid-template-rows: 10px 20px,
        grid-template-columns: 10% 20%,
        grid-template-areas: none,
        grid-lanes-direction: row should be valid. assert_equals: expected (string) "10px 20px row" but got (undefined) undefined
FAIL grid-template-rows: none,
        grid-template-columns: 1fr 1fr 3fr,
        grid-template-areas: "a a b",
        grid-lanes-direction: column should be valid. assert_equals: expected (string) "\"a a b\" 1fr 1fr 3fr column" but got (undefined) undefined
FAIL grid-template-rows: 20% 40%,
        grid-template-columns: none,
        grid-template-areas: "b" "a",
        grid-lanes-direction: row should be valid. assert_equals: expected (string) "\"b a\" 20% 40% row" but got (undefined) undefined
FAIL grid-template-rows: none,
        grid-template-columns: fit-content(calc(0.5em + 10px)),
        grid-template-areas: none,
        grid-lanes-direction: column track-reverse fill-reverse should be valid. assert_equals: expected (string) "fit-content(calc(0.5em + 10px)) column track-reverse fill-reverse" but got (undefined) undefined
FAIL grid-template-rows: 10% 20% 40%,
        grid-template-columns: none,
        grid-template-areas: "a" "b" "c",
        grid-lanes-direction: row fill-reverse track-reverse should be valid. assert_equals: expected (string) "\"a b c\" 10% 20% 40% row fill-reverse track-reverse" but got (undefined) undefined
FAIL grid-template-rows: repeat(2, 1fr),
        grid-template-columns: 10px 20px,
        grid-template-areas: "a b",
        grid-lanes-direction: normal should be valid. assert_equals: expected (string) "\"a b\" 10px 20px normal" but got (undefined) undefined
FAIL grid-lanes: normal 100px, then grid-template-rows: 200px assert_equals: grid-lanes should serialize as: 100px normal expected (string) "100px normal" but got (undefined) undefined
FAIL grid-lanes: normal 100px, then grid-template-columns: 200px assert_equals: grid-lanes should serialize as: 200px normal expected (string) "200px normal" but got (undefined) undefined
FAIL grid-lanes: row 100px, then grid-lanes-direction: normal assert_equals: grid-lanes should serialize as: normal expected (string) "normal" but got (undefined) undefined
FAIL grid-lanes: column 100px, then grid-lanes-direction: normal assert_equals: grid-lanes should serialize as: 100px normal expected (string) "100px normal" but got (undefined) undefined
FAIL grid-lanes-direction: column, then grid-lanes: row 100px 200px assert_equals: grid-lanes should serialize as: 100px 200px row expected (string) "100px 200px row" but got (undefined) undefined
FAIL grid-lanes: "a b" 100px column, then grid-template-areas: "c d" assert_equals: grid-lanes should serialize as: "c d" 100px column expected (string) "\"c d\" 100px column" but got (undefined) undefined
FAIL grid-template-columns: 10px, grid-lanes-direction: column, then grid-lanes: row 100px assert_equals: grid-lanes should serialize as: 100px row expected (string) "100px row" but got (undefined) undefined
